+++ src/renderer/components/index.ts
@@ -1,0 +1,4 @@
+export * from './main.component';
+export * from './settings.component';
+export * from './home.component';
+export * from './steam-devices.component';