+++ src/renderer/services/index.ts
@@ -1,0 +1,3 @@
+export * from './settings.service';
+export * from './steam-devices.service';
+export * from './window.service';